GoRouterの設定をMyAppに付加して、MyHomePageを別ファイルに移動
