+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:yumemi_flutter_engineer_codecheck/l10n/app_localizations.dart';
 import 'package:yumemi_flutter_engineer_codecheck/provider/custom_theme_data_provider.dart';
+import 'package:yumemi_flutter_engineer_codecheck/provider/go_router_provider.dart';
 import 'package:yumemi_flutter_engineer_codecheck/provider/selected_theme_mode_provider.dart';
-import 'package:yumemi_flutter_engineer_codecheck/view/widget/theme_mode_select_button.dart';
 
 void main() {
   runApp(const ProviderScope(child: MyApp()));
@@ -15,61 +15,14 @@ class MyApp extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final themeMode = ref.watch(selectedThemeModeProvider);
-    return MaterialApp(
+    return MaterialApp.router(
       title: 'Flutter Demo',
       theme: ref.watch(customThemeDataProvider(Brightness.light)),
       darkTheme: ref.watch(customThemeDataProvider(Brightness.dark)),
       themeMode: themeMode.value ?? ThemeMode.system,
       localizationsDelegates: AppLocalizations.localizationsDelegates,
       supportedLocales: AppLocalizations.supportedLocales,
-      home: const MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key});
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  var _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final localizations = AppLocalizations.of(context)!;
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(localizations.homePageTitle),
-        actions: const [
-          ThemeModeSelectButton(),
-        ],
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(localizations.counterText),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: localizations.incrementTooltip,
-        child: const Icon(Icons.add),
-      ),
+      routerConfig: ref.watch(goRouterProvider),
     );
   }
 }
